Use StandardRb instead of RuboCop
Judging by the length and history of .rubocop_todo.yml, we clearly don't have a
lot of preference for one style or another in this repository. I suggest that
we just use StandardRb, and fix all the code it's objecting to. There's more to
life than talking about Ruby code linters (or so I've heard).