Add `specs_passing?` job
We previously had a required check that Rails 7.0 passed with Ruby 3.2.
We no longer have that specific run so I had to remove the check.

This adds a `specs_successful?` check that will return `true` if the
entire matrix for `specs` passes. We can use this job as the required
check instead to avoid having to update the checks every time the matrix changes.